Revert "[REFACTOR] Move HTML-STF and IntelliJ-STF parts to STF-Project (#216)"
This reverts commit de080fe248a94cc9b378b1982b6fdb79e518ec8a.
