+++ de.fu_berlin.inf.dpp.ui/META-INF/MANIFEST.MF
@@ -13,9 +13,13 @@ Bundle-ClassPath: .,
 Export-Package: de.fu_berlin.inf.ag_se.browser,
  de.fu_berlin.inf.ag_se.browser.extensions,
  de.fu_berlin.inf.ag_se.browser.functions,
- de.fu_berlin.inf.ag_se.browser.html,
  de.fu_berlin.inf.ag_se.browser.swt,
  de.fu_berlin.inf.dpp,
+ de.fu_berlin.inf.dpp.stf.server,
+ de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot,
+ de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.impl,
+ de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget,
+ de.fu_berlin.inf.dpp.stf.server.rmi.htmlbot.widget.impl,
  de.fu_berlin.inf.dpp.ui,
  de.fu_berlin.inf.dpp.ui.browser_functions,
  de.fu_berlin.inf.dpp.ui.ide_embedding,
