+++ de.fu_berlin.inf.dpp.ui/build.properties
@@ -1,4 +1,5 @@
-source.. = src/,
+source.. = src/,\
+           test/framework/stf/src/
 output.. = bin/
 bin.includes = META-INF/,\
                .,\
